move chatgpt helper into own file
